Generalised the Particle constructor and created a new more specific constructor for a gas particle called GasParticle.
I hope that eventually index.html will eventually contain multiple different simulations which can share the same basic implementation of a particle. Therefore it made sense for me to generalise my implementation of Particle. To this end, any functionality which feels specific to a gas particle (such as all particles in my simulation having the same radius and mass) will be put in the GasParticle constructor instead.